requirements.txt
async-timeout==5.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
oauthlib==3.3.1
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
tweepy==4.16.0
urllib3==2.5.0
